requirements.txt
pandas~=2.2.3
openpyxl~=3.1.5
curl_cffi~=0.7.4
gspread~=6.1.4
python-dotenv~=1.0.1
ratelimit~=2.2.1
tabulate~=0.9.0